Convert binary to bundle pattern
