--- cmd/workflow-engine/BUILD.bazel
@@ -1,22 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["workflow-engine.go"],
-    visibility = ["//visibility:private"],
-    deps = [
-        "//cmd/workflow-engine/app:go_default_library",
-        "//pkg/api/function:go_default_library",
-        "//pkg/fnenv/native:go_default_library",
-        "//pkg/fnenv/native/builtin:go_default_library",
-        "//vendor/github.com/nats-io/go-nats:go_default_library",
-        "//vendor/github.com/sirupsen/logrus:go_default_library",
-        "//vendor/github.com/urfave/cli:go_default_library",
-    ],
-)
-
-go_binary(
-    name = "workflow-engine",
-    library = ":go_default_library",
-    visibility = ["//visibility:public"],
-)
