--- cmd/workflow-engine/app/BUILD.bazel
@@ -1,49 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["bootstrap.go"],
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/api/function:go_default_library",
-        "//pkg/api/invocation:go_default_library",
-        "//pkg/api/workflow:go_default_library",
-        "//pkg/api/workflow/parse:go_default_library",
-        "//pkg/apiserver:go_default_library",
-        "//pkg/controller:go_default_library",
-        "//pkg/controller/query:go_default_library",
-        "//pkg/fes:go_default_library",
-        "//pkg/fes/eventstore/nats:go_default_library",
-        "//pkg/fnenv/fission:go_default_library",
-        "//pkg/scheduler:go_default_library",
-        "//pkg/types/aggregates:go_default_library",
-        "//pkg/types/typedvalues:go_default_library",
-        "//pkg/util/labels:go_default_library",
-        "//pkg/util/pubsub:go_default_library",
-        "//vendor/github.com/gorilla/handlers:go_default_library",
-        "//vendor/github.com/grpc-ecosystem/grpc-gateway/runtime:go_default_library",
-        "//vendor/github.com/nats-io/go-nats-streaming:go_default_library",
-        "//vendor/github.com/sirupsen/logrus:go_default_library",
-        "//vendor/google.golang.org/grpc:go_default_library",
-    ],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["bootstrap_test.go"],
-    library = ":go_default_library",
-    deps = [
-        "//pkg/api/function:go_default_library",
-        "//pkg/apiserver:go_default_library",
-        "//pkg/fnenv/native/builtin:go_default_library",
-        "//pkg/fnenv/test:go_default_library",
-        "//pkg/types:go_default_library",
-        "//pkg/types/typedvalues:go_default_library",
-        "//vendor/github.com/golang/protobuf/ptypes/empty:go_default_library",
-        "//vendor/github.com/nats-io/go-nats:go_default_library",
-        "//vendor/github.com/nats-io/go-nats-streaming:go_default_library",
-        "//vendor/github.com/sirupsen/logrus:go_default_library",
-        "//vendor/github.com/stretchr/testify/assert:go_default_library",
-        "//vendor/google.golang.org/grpc:go_default_library",
-    ],
-)
